+++ fragments.js
@@ -1,14 +1,12 @@
 import { graphql } from 'gatsby'
 
-export const GatsbyPrismicImageFixedType = graphql`
+export const GatsbyPrismicImageFixed = graphql`
   fragment GatsbyPrismicImageFixed on PrismicImageFixedType {
     base64
     width
     height
     src
     srcSet
-    srcWebp
-    srcSetWebp
   }
 `
 
@@ -18,13 +16,11 @@ export const GatsbyPrismicImageFixed_noBase64 = graphql`
     height
     src
     srcSet
-    srcWebp
-    srcSetWebp
   }
 `
 
-// Not actually necessary - since Imgix is scaling,
-// there is no "penalty" for including WebP by default
+// Not actually necessary since applying `auto=format` to the source URL
+// automatically returns WebP for all URLs.
 export const GatsbyPrismicImageFixed_withWebp = graphql`
   fragment GatsbyPrismicImageFixed_withWebp on PrismicImageFixedType {
     base64
@@ -37,8 +33,8 @@ export const GatsbyPrismicImageFixed_withWebp = graphql`
   }
 `
 
-// Not actually necessary - since Imgix is scaling,
-// there is no "penalty" for including WebP by default
+// Not actually necessary since applying `auto=format` to the source URL
+// automatically returns WebP for all URLs.
 export const GatsbyPrismicImageFixed_withWebp_noBase64 = graphql`
   fragment GatsbyPrismicImageFixed_withWebp_noBase64 on PrismicImageFixedType {
     width
@@ -50,14 +46,12 @@ export const GatsbyPrismicImageFixed_withWebp_noBase64 = graphql`
   }
 `
 
-export const GatsbyPrismicImageFluidType = graphql`
+export const GatsbyPrismicImageFluid = graphql`
   fragment GatsbyPrismicImageFluid on PrismicImageFluidType {
     base64
     aspectRatio
     src
     srcSet
-    srcWebp
-    srcSetWebp
     sizes
   }
 `
@@ -67,8 +61,6 @@ export const GatsbyPrismicImageFluid_noBase64 = graphql`
     aspectRatio
     src
     srcSet
-    srcWebp
-    srcSetWebp
     sizes
   }
 `