+++ src/documentsToNodes.ts
@@ -1,4 +1,3 @@
-import pascalcase from 'pascalcase'
 import pick from 'lodash.pick'
 import omit from 'lodash.omit'
 
@@ -30,7 +29,7 @@ const getTypeForPath = (
   typePaths: TypePath[],
 ): GraphQLType | string | undefined => {
   const stringifiedPath = JSON.stringify(path)
-  const def = typePaths.find(x => JSON.stringify(x.path) === stringifiedPath)
+  const def = typePaths.find((x) => JSON.stringify(x.path) === stringifiedPath)
 
   if (!def) return
   if (/^\[.*GroupType\]$/.test(def.type)) return GraphQLType.Group
@@ -68,7 +67,8 @@ const normalizeField = async (
 
       const base = await normalizeImageField(apiId, baseObj, path, doc, env)
       const thumbs = await mapObjValsP(
-        async thumb => await normalizeImageField(apiId, thumb, path, doc, env),
+        async (thumb) =>
+          await normalizeImageField(apiId, thumb, path, doc, env),
         thumbsObj,
       )
 
@@ -126,8 +126,8 @@ const normalizeField = async (
             primary: normalizedPrimary,
             items: normalizedItems,
             internal: {
-              type: pascalcase(
-                `Prismic ${doc.type} ${apiId} ${slice.slice_type}`,
+              type: buildSchemaTypeName(
+                `${doc.type} ${apiId} ${slice.slice_type}`,
               ),
               contentDigest: createContentDigest(slice),
             },
@@ -156,7 +156,7 @@ const normalizeField = async (
       // there by default.
       return await Promise.all(
         (field as AlternateLanguagesField).map(
-          async item =>
+          async (item) =>
             await normalizeLinkField(
               apiId,
               {
@@ -193,7 +193,7 @@ const normalizeObjs = (
   path: TypePath['path'],
   doc: PrismicDocument,
   env: DocumentsToNodesEnvironment,
-) => Promise.all(objs.map(obj => normalizeObj(obj, path, doc, env)))
+) => Promise.all(objs.map((obj) => normalizeObj(obj, path, doc, env)))
 
 export const documentToNodes = async (
   doc: PrismicDocument,
@@ -206,7 +206,7 @@ export const documentToNodes = async (
   if (linkResolver) linkResolverForDoc = linkResolver({ node: doc })
 
   const docNodeId = createNodeId(`${doc.type} ${doc.id}`)
-  const docUrl = linkResolverForDoc?.(doc)
+  const docUrl = linkResolverForDoc ? linkResolverForDoc(doc) : undefined
 
   const normalizedData = await normalizeObj(
     doc.data,
@@ -245,4 +245,4 @@ export const documentToNodes = async (
 export const documentsToNodes = async (
   docs: PrismicDocument[],
   env: DocumentsToNodesEnvironment,
-) => await Promise.all(docs.map(doc => documentToNodes(doc, env)))
+) => await Promise.all(docs.map((doc) => documentToNodes(doc, env)))