+++ src/index.ts
@@ -1,3 +1,9 @@
 export { usePrismicPreview } from './usePrismicPreview'
 export { mergePrismicPreviewData } from './mergePrismicPreviewData'
-export { buildFixedGatsbyImage, buildFluidGatsbyImage } from './gatsbyImage'
+export { withPreview } from './withPreview'
+export {
+  withPreviewResolver,
+  WithPreviewResolverProps,
+} from './withPreviewResolver'
+export { withUnpublishedPreview } from './withUnpublishedPreview'
+export { usePreviewStore, PreviewStoreProvider } from './usePreviewStore'