+++ src/utils.ts
@@ -1,6 +1,5 @@
-import pascalcase from 'pascalcase'
-
-import { name as pkgName } from '../package.json'
+import * as React from 'react'
+import { pascalCase, pascalCaseTransformMerge } from 'pascal-case'
 
 /**
  * Returns a namespaced string intended for logging.
@@ -9,7 +8,7 @@ import { name as pkgName } from '../package.json'
  *
  * @returns Namespaced message.
  */
-export const msg = (message: string) => `${pkgName} - ${message}`
+export const msg = (message: string) => `gatsby-source-prismic - ${message}`
 
 /**
  * Maps key-value tuples of an object to new key-value tuples to create a new
@@ -25,7 +24,7 @@ export const mapObj = <T1, T2>(
   obj: { [key: string]: T1 },
 ): { [key: string]: T2 } => {
   const entries = Object.entries(obj)
-  const pairs = entries.map(x => fn(x))
+  const pairs = entries.map((x) => fn(x))
 
   const result: { [key: string]: T2 } = {}
 
@@ -51,7 +50,7 @@ export const mapObjP = async <T1, T2>(
   obj: { [key: string]: T1 },
 ): Promise<{ [key: string]: T2 }> => {
   const entries = Object.entries(obj)
-  const pairs = await Promise.all(entries.map(x => Promise.resolve(fn(x))))
+  const pairs = await Promise.all(entries.map((x) => Promise.resolve(fn(x))))
 
   const result: { [key: string]: T2 } = {}
 
@@ -98,7 +97,7 @@ export const mapObjValsP = async <T1, T2>(
 
   const keys = Object.keys(obj)
   await Promise.all(
-    keys.map(async key => {
+    keys.map(async (key) => {
       result[key] = await fn(obj[key], key)
     }),
   )
@@ -126,7 +125,7 @@ export const isEmptyObj = (obj: object) => {
  * @returns Type name for the schema.
  */
 export const buildSchemaTypeName = (apiId: string) =>
-  pascalcase(`Prismic ${apiId}`)
+  pascalCase(`Prismic ${apiId}`, { transform: pascalCaseTransformMerge })
 
 /**
  * Determines whether the current context is the browser.
@@ -134,3 +133,7 @@ export const buildSchemaTypeName = (apiId: string) =>
  * @returns `true` if the current context is the browser, `false` otherwise.
  */
 export const isBrowser = typeof window !== 'undefined'
+
+export const getComponentDisplayName = (
+  WrappedComponent: React.ComponentType<any>,
+) => WrappedComponent.displayName || WrappedComponent.name || 'Component'