+++ src/validateOptions.ts
@@ -5,12 +5,24 @@ import { PluginOptions, BrowserPluginOptions } from './types'
 const baseSchema = {
   repositoryName: 'string',
   accessToken: 'string?',
+  releaseID: 'string?',
+  schemas: struct.record(['string', 'object']),
   linkResolver: 'function?',
   htmlSerializer: 'function?',
   fetchLinks: struct.optional(['string']),
   lang: 'string?',
   typePathsFilenamePrefix: 'string?',
-  prismicToolbar: 'boolean?',
+  prismicToolbar: struct.optional(
+    struct.union(['boolean', struct.enum(['legacy'])]),
+  ),
+  imageImgixParams: struct.record([
+    'string',
+    struct.union(['string', 'number', 'boolean', 'undefined']),
+  ]),
+  imagePlaceholderImgixParams: struct.record([
+    'string',
+    struct.union(['string', 'number', 'boolean', 'undefined']),
+  ]),
 }
 
 const baseDefaults = {
@@ -20,12 +32,21 @@ const baseDefaults = {
   lang: '*',
   typePathsFilenamePrefix: 'prismic-typepaths---',
   prismicToolbar: false,
+  imageImgixParams: {
+    auto: 'format,compress',
+    fit: 'max',
+    q: 50,
+  },
+  imagePlaceholderImgixParams: {
+    w: 100,
+    blur: 15,
+    q: 20,
+  },
 }
 
 const PluginOptionsValidator = struct(
   {
     ...baseSchema,
-    schemas: struct.record(['string', 'object']),
     shouldDownloadImage: 'function?',
     plugins: struct.size([0, 0]),
   },