+++ test/gatsby-node.test.ts
@@ -1,17 +1,11 @@
-import {
-  SourceNodesArgs,
-  CreateResolversArgs,
-  ParentSpanPluginArgs,
-} from 'gatsby'
+import fs from 'fs'
+import { SourceNodesArgs, ParentSpanPluginArgs } from 'gatsby'
 
-import {
-  sourceNodes,
-  createResolvers,
-  onPreExtractQueries,
-} from '../src/gatsby-node'
+import { sourceNodes, onPreExtractQueries } from '../src/gatsby-node'
 
-import { writeFileSync, copyFileSync } from './__mocks__/fs-extra'
 import mockSchema from './__fixtures__/schema.json'
+
+jest.mock('fs')
 
 const PROGRAM_DIRECTORY_PATH = '/__PROGRAM_DIRECTORY__/'
 
@@ -102,14 +96,8 @@ const mockGatsbyContext: ParentSpanPluginArgs = {
   },
   getNodeAndSavePathDependency: jest.fn(),
   cache: {
-    getAndPassUp: jest.fn(),
-    wrap: jest.fn(),
     set: jest.fn(),
-    mset: jest.fn(),
     get: jest.fn(),
-    mget: jest.fn(),
-    del: jest.fn(),
-    reset: jest.fn(),
   },
   createNodeId: jest.fn().mockReturnValue('createNodeId'),
   createContentDigest: jest.fn().mockReturnValue('createContentDigest'),
@@ -122,23 +110,23 @@ const mockGatsbyContext: ParentSpanPluginArgs = {
   schema: {
     buildObjectType: jest
       .fn()
-      .mockImplementation(config => ({ kind: 'OBJECT', config })),
-    buildUnionType: jest.fn().mockImplementation(config => ({
+      .mockImplementation((config) => ({ kind: 'OBJECT', config })),
+    buildUnionType: jest.fn().mockImplementation((config) => ({
       kind: 'UNION',
       config,
     })),
     buildInterfaceType: jest
       .fn()
-      .mockImplementation(config => ({ kind: 'INTERFACE', config })),
+      .mockImplementation((config) => ({ kind: 'INTERFACE', config })),
     buildInputObjectType: jest
       .fn()
-      .mockImplementation(config => ({ kind: 'INPUT', config })),
+      .mockImplementation((config) => ({ kind: 'INPUT', config })),
     buildEnumType: jest
       .fn()
-      .mockImplementation(config => ({ kind: 'ENUM', config })),
+      .mockImplementation((config) => ({ kind: 'ENUM', config })),
     buildScalarType: jest
       .fn()
-      .mockImplementation(config => ({ kind: 'SCALAR', config })),
+      .mockImplementation((config) => ({ kind: 'SCALAR', config })),
   },
 }
 
@@ -146,13 +134,6 @@ const mockSourceNodesGatsbyContext: SourceNodesArgs = {
   ...mockGatsbyContext,
   traceId: 'initial-sourceNodes',
   waitForCascadingActions: false,
-}
-
-const mockCreateResolversGatsbyContext: CreateResolversArgs = {
-  ...mockSourceNodesGatsbyContext,
-  traceId: 'initial-createResolvers',
-  intermediateSchema: {},
-  createResolvers: jest.fn(),
 }
 
 const pluginOptions = {
@@ -179,31 +160,16 @@ describe('sourceNodes', () => {
   test('writes type paths to filesystem', async () => {
     await sourceNodes!(mockSourceNodesGatsbyContext, pluginOptions)
 
-    expect(writeFileSync.mock.calls[0][0]).toMatchInlineSnapshot(
+    expect(
+      (fs.writeFileSync as jest.Mock).mock.calls[0][0],
+    ).toMatchInlineSnapshot(
       `"/__PROGRAM_DIRECTORY__/public/prismic-typepaths---9769f52526da286b236e9bd2cb0d0291.json"`,
     )
 
     // Ensure valid JSON.
-    expect(JSON.parse(writeFileSync.mock.calls[0][1])).toMatchSnapshot()
-  })
-})
-
-describe('createResolvers', () => {
-  test('creates resolvers', async () => {
-    await createResolvers!(mockCreateResolversGatsbyContext, pluginOptions)
-
     expect(
-      mockCreateResolversGatsbyContext.createResolvers,
-    ).toHaveBeenCalledWith({
-      PrismicImageType: {
-        fixed: { resolve: expect.any(Function) },
-        fluid: { resolve: expect.any(Function) },
-      },
-      PrismicImageThumbnailType: {
-        fixed: { resolve: expect.any(Function) },
-        fluid: { resolve: expect.any(Function) },
-      },
-    })
+      JSON.parse((fs.writeFileSync as jest.Mock).mock.calls[0][1]),
+    ).toMatchSnapshot()
   })
 })
 
@@ -211,7 +177,7 @@ describe('onPreExtractQueries', () => {
   test('copies fragments file to program cache', async () => {
     onPreExtractQueries!(mockGatsbyContext, pluginOptions)
 
-    const call = copyFileSync.mock.calls[0]
+    const call = (fs.copyFileSync as jest.Mock).mock.calls[0]
 
     expect(call[0]).toMatch(/\/fragments.js$/)
     expect(call[1]).toBe(